pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ads-grafana-toolkit"
version = "0.1.0"
description = "A toolkit for simplifying Grafana dashboard creation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
keywords = ["grafana", "dashboard", "monitoring", "prometheus", "metrics"]
dependencies = [
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "click>=8.0",
    "rich>=13.0",
    "questionary>=2.0",
]

[project.optional-dependencies]
web = [
    "flask>=3.0",
]
nlp = [
    "openai>=1.0",
]
all = [
    "flask>=3.0",
    "openai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "flask>=3.0",
]

[project.scripts]
ads-grafana-toolkit = "ads_grafana_toolkit.cli.main:cli"
agt = "ads_grafana_toolkit.cli.main:cli"

[project.urls]
Homepage = "https://github.com/afterdarktech/ads-grafana-toolkit"
Repository = "https://github.com/afterdarktech/ads-grafana-toolkit"

[tool.setuptools.packages.find]
where = ["."]
include = ["ads_grafana_toolkit*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
